Extract blog template into plugin
