+++ src/plugins/blog-posts/index.js
@@ -1,0 +1,1 @@
+// no-op
