+++ src/plugins/blog-posts/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "blog-posts",
+  "private": true,
+  "version": "0.0.0-dev",
+  "main": "index.js"
+}
